fix: Fix Windows installation on fresh device (#6)
When the service has never been installed before, the virtual service
account doesn't exist yet. So we must install the service before setting
the ACL on the runtime folder.

This was overlooked because on my development machine, I had done things
in a different order during the initial development of the Windows
package.