Phase 4a: Add shaka-lab-gateway packages for Linux (#13)
The gateway package depends on and configures a suite of network
services for DHCP, DNS, and Active Directory Domain Controller for the
lab network. It can also configure the device as a NAT router
(optional).

The gateway client package configures devices in the lab to allow
centrally-managed Active Directory users to log in.

I plan to add client packages to configure Mac & Windows later.