fix: Fix Active Directory connections for Windows clients (#26)
I was unable add a Windows client to the domain created by
`shaka-lab-gateway`. This set of changes fixes the issue.

- Align the DNS domain name with the AD domain name (`lab.shaka`). This
is generally recommended for AD.
- Serve SRV records directly in DNS server rather than have Samba inject
them into the DNS server dynamically. (In fact, the dynamic injection
fails when you have already defined a domain, which is why we had
different DNS and AD domains in the first place.)