fix: Convert shaka-lab-node from Formula into Cask (#28)
This simplifies the Java dependency and allows us to automatically
restart services, since Cask installations are not sandboxed like
Formulae. The lack of sandboxing also allows us to simplify the
installation paths somewhat.

Now all macOS packages are Homebrew Casks, not Formulae, which
simplifies the deployment to use wildcards instead of explicit package
lists.

Finally, I discovered that we don't need the --cask flag and that we can
explicitly update individual packages instead of all-at-once, so I've
udpated all the macOS package docs with these details.