fix: Fix macOS Selenium node WRT autologin
On macOS, headless mode fails to take screenshots, so a GUI session is
required.  The best way to manage this is with a user set for
autologin, and the node service running as this user.

To streamline this, shaka-lab-recommended-settings now configures the
system for autologin for the user installing that package, and
shaka-lab-node runs as the logged in GUI user.

Once installed, the shaka-lab-node service can now be managed by a
user other than the GUI user.  This allows us to have a non-personal
account for testing, but individual user accounts for each admin to
manage the system via SSH.

This also:
 - Updates documentation that was missed when we transitioned from
   Homebrew Formula to Cask.  The Cask can restart services for you,
   so you don't need to do that yourself on install or upgrade.
 - Updates documentation to use sudo on service scripts
   (restart/stop), which allows us to perform these actions even if
   the admin account differs from the logged in GUI account used for
   testing.
 - Moves installation from a variable path (relative to Homebrew) to a
   fixed path (/opt/shaka-lab-node), since everything is managed
   through sudo now and the install path doesn't have to be directly
   writable by the admin doing upgrades or making changes.