fix: Avoid Firefox snap installation
An automated upgrade on one of our lab devices reverted Firefox to the
"official" Ubuntu snap package, overriding our choice of the Mozilla
PPA version.

This ensures that automatic upgrades can use the PPA version as well,
and creates a versioned dependency in shaka-lab-browsers so that only
the PPA version will be accepted there.

Affected machines will need to uninstall firefox and
shaka-lab-browsers, then reinstall.