fix(Windows): Fix hung drivers on Windows (#44)
These hung drivers, when they occur, prevent the executable files from being deleted or updated. We have to kill common drivers before attempting to update. With this, the system can recover after updating shaka-lab-node.

We also needed to add a check to the main service to fail when a driver update fails for any reason.