feat(shaka-lab-node): Add support for Chromecast redirect mode (#52)
Chromecast WebDriver Server v2 adds support for a redirect mode, where
it does not rely on an iframe. This has many advantages for Shaka Player
testing.

This adds support for this mode in shaka-lab-node, and enables it by
default in the Shaka lab.

To accomplish this for Chromecast, we add general support for optional
parameters with default values in the node template.