fix(shaka-lab-github-runner): Fix service restart logic (#59)
We should always try to restart these runners, but the systemd config
for that was unclear. This fixes the service logic to never give up on
restarting the service.

We had a transient error in authentication to GitHub that took down
15/16 runner services. Once they reached their start limit, they stopped
completely in a failed state.

With this configuration, that would not have happened. Eventually, they
all would have come back up.