fix(shaka-lab-node): Refactor services for macos 15 (#78)
macOS 15 introduced local network permissions that broke shaka-lab-node
services. These should in theory be fixed by accepting a permissions
prompt, but the prompt is never shown for launchd agents. This has been
widely discussed in Apple forums, with many people complaining
specifically about how this is breaking CI systems. So we are not alone.

The best workaround seems to be building our background services and
other tasks into Automator apps. Those can be added as login items, and
they seem to be able to get the appropriate permissions if they live in
`/Applications`.

The workflow sources are in
`shaka-lab-node/macos/*.app/Contents/document.wflow`.

It's unfortunate that we have to check in the entire `.app` bundle,
including binaries, instead of building it on the fly from the script
inside `document.wflow`. I tried, but I was unable to figure out how to
do this entirely by using AppleScript to direct Automator. (I got close,
but there's a prompt I can't figure out how to get past in an automated
way.)